+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Added
 
 - added the ability to increase the speed of uploading files to ESP. The baud rate is set in the `uploadBaudrate` setting.
+- added folder deletion on the device
 
 ## [5.1.0] - 2025-01-26