+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [5.2.0] - 2025-03-01
+
 ## Added
 
 - added the ability to increase the speed of uploading files to ESP. The baud rate is set in the `uploadBaudrate` setting.